requirements.txt
hidapi>=0.14
PySide6>=6.6
Pillow>=10.0
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
pynput>=1.7.6; sys_platform == "linux"
python-xlib>=0.33; sys_platform == "linux"
evdev>=1.6.0; sys_platform == "linux"
